added authorization for some user credential routes

+++ controllers/api/user-profile-routes.js
@@ -1,5 +1,6 @@
 const router = require("express").Router();
 const { User } = require("../../models");
+const {withAuth} = require("../../utils/withAuth");
 
 // Already have route to get all users
 router.get("/", (req, res) => {});
@@ -17,7 +18,7 @@ router.get("/:id", async (req, res) => {
 router.post("/", (req, res) => {});
 
 // authorize route I think
-router.put("/:id", async (req, res) => {
+router.put("/:id", withAuth, async (req, res) => {
     User.update(req.body, {
         where: {
             id: req.params.id,
@@ -30,17 +31,23 @@ router.put("/:id", async (req, res) => {
         });
 });
 
-// Definitely authorize this
-router.delete("/:id", async (req, res) => {
-    try {
-        const userData = await User.destroy({ where: { id: req.params.id }});
-        if(!userData){
-            res.status(404).json({ message: "No user found with that ID" });
-            return;
+// Delete an existing user
+// Can only delete the user they're logged in as
+// ** ADD CODE TO LOG THEM OUT ONCE DELETED IF NEEDED **
+router.delete("/:id", withAuth, async (req, res) => {
+    if(req.session.user_id === req.params.id){
+        try {
+            const userData = await User.destroy({ where: { id: req.params.id }});
+            if(!userData){
+                res.status(404).json({ message: "No user found with that ID" });
+                return;
+            }
+            res.status(200).json(userData);
+        } catch (err) {
+            res.status(500).json(err);
         }
-        res.status(200).json(userData);
-    } catch (err) {
-        res.status(500).json(err);
+    } else {
+        res.status(403).json({ message: "You are not logged in as the user you are attempting to delete." });
     }
 });
